+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/PayrollPluginConst.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BTCPayServer.RockstarDev.Plugins.Payroll
+﻿namespace BTCPayServer.RockstarDev.Plugins.Payroll
 {
     // Need to extract these variables, their references are places for improvements
     public class PayrollPluginConst
