+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Views/PayrollInvoice/Upload.cshtml
@@ -4,7 +4,7 @@
 @using BTCPayServer.Client
 @using BTCPayServer.RockstarDev.Plugins.Payroll.Data.Models
 @using BTCPayServer.Services
-@using static BTCPayServer.RockstarDev.Plugins.Payroll.PayrollInvoiceController
+@using static BTCPayServer.RockstarDev.Plugins.Payroll.Controllers.PayrollInvoiceController
 @inject IScopeProvider ScopeProvider
 @inject DisplayFormatter DisplayFormatter
 @model PayrollInvoiceUploadViewModel
