+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Views/PayrollUser/Create.cshtml
@@ -4,7 +4,7 @@
 @using BTCPayServer.Client
 @using BTCPayServer.RockstarDev.Plugins.Payroll.Data.Models
 @using BTCPayServer.Services
-@using static BTCPayServer.RockstarDev.Plugins.Payroll.PayrollUserController
+@using static BTCPayServer.RockstarDev.Plugins.Payroll.Controllers.PayrollUserController
 @inject IScopeProvider ScopeProvider
 @inject DisplayFormatter DisplayFormatter
 @model PayrollUserCreateViewModel
