+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Views/Public/ChangePassword.cshtml
@@ -1,10 +1,4 @@
-﻿@using BTCPayServer.Abstractions.Extensions
-@using BTCPayServer.Abstractions.Contracts
-@using BTCPayServer.Abstractions.Models
-@using BTCPayServer.Client
-@using BTCPayServer.RockstarDev.Plugins.Payroll
-@using BTCPayServer.RockstarDev.Plugins.Payroll.Data.Models
-@using BTCPayServer.Services
+﻿@using BTCPayServer.RockstarDev.Plugins.Payroll.Controllers
 @model PublicChangePasswordViewModel
 @{
     Layout = "_PayrollPublicLayout";
