+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Views/Public/ListInvoices.cshtml
@@ -1,6 +1,6 @@
 ﻿@using BTCPayServer.Abstractions.Contracts
 @using BTCPayServer.Abstractions.Extensions
-@using BTCPayServer.RockstarDev.Plugins.Payroll
+@using BTCPayServer.RockstarDev.Plugins.Payroll.Controllers
 @inject BTCPayServer.Services.BTCPayServerEnvironment Env
 @model PublicListInvoicesViewModel
 @{
