+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Views/Public/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@using BTCPayServer.RockstarDev.Plugins.Payroll
+﻿@using BTCPayServer.RockstarDev.Plugins.Payroll.Controllers
 @inject BTCPayServer.Services.BTCPayServerEnvironment Env
 @model PublicLoginViewModel
 @{
