+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Views/Public/_PayrollPublicLayout.cshtml
@@ -1,5 +1,5 @@
-@using BTCPayServer.RockstarDev.Plugins.Payroll
-@using static BTCPayServer.RockstarDev.Plugins.Payroll.PublicController
+@using BTCPayServer.RockstarDev.Plugins.Payroll.Controllers
+@using static BTCPayServer.RockstarDev.Plugins.Payroll.Controllers.PublicController
 @inject BTCPayServer.Services.BTCPayServerEnvironment Env
 @model BasePayrollPublicViewModel
 @{
